+++ examples/data-sources/nxos_ospf_area/data-source.tf
@@ -1,5 +1,5 @@
 data "nxos_ospf_area" "example" {
   instance_name = "OSPF1"
-  vrf_name      = "VRF1"
+  vrf_name      = "default"
   area_id       = "0.0.0.10"
 }