+++ examples/data-sources/nxos_ospf_vrf/data-source.tf
@@ -1,4 +1,4 @@
 data "nxos_ospf_vrf" "example" {
   instance_name = "OSPF1"
-  name          = "VRF1"
+  name          = "default"
 }