+++ examples/resources/nxos_ospf_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ospf_interface.example "sys/ospf/inst-[OSPF1]/dom-[default]/if-[eth1/10]"