+++ examples/resources/nxos_ospf_vrf/import.sh
@@ -1,1 +1,1 @@
-terraform import nxos_ospf_vrf.example "sys/ospf/inst-[OSPF1]/dom-[VRF1]"
+terraform import nxos_ospf_vrf.example "sys/ospf/inst-[OSPF1]/dom-[default]"