+++ internal/provider/data_source_nxos_ospf_area_test.go
@@ -41,7 +41,7 @@ func TestAccDataSourceNxosOSPFArea(t *testing.T) {
 const testAccDataSourceNxosOSPFAreaConfig = `
 data "nxos_ospf_area" "test" {
   instance_name = "OSPF1"
-  vrf_name = "VRF1"
+  vrf_name = "default"
   area_id = "0.0.0.10"
 }
 `