+++ internal/provider/helpers/utils.go
@@ -1,0 +1,8 @@
+package helpers
+
+func ParseNxosBoolean(s string) bool {
+	if s == "yes" || s == "true" {
+		return true
+	}
+	return false
+}