+++ internal/provider/provider.go
@@ -246,6 +246,7 @@ func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceT
 		"nxos_ospf_instance":          resourceOSPFInstanceType{},
 		"nxos_ospf_vrf":               resourceOSPFVRFType{},
 		"nxos_ospf_area":              resourceOSPFAreaType{},
+		"nxos_ospf_interface":         resourceOSPFInterfaceType{},
 	}, nil
 }
 
@@ -277,6 +278,7 @@ func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSou
 		"nxos_ospf_instance":          dataSourceOSPFInstanceType{},
 		"nxos_ospf_vrf":               dataSourceOSPFVRFType{},
 		"nxos_ospf_area":              dataSourceOSPFAreaType{},
+		"nxos_ospf_interface":         dataSourceOSPFInterfaceType{},
 	}, nil
 }